Do not use a default callbacks directory
Callback usage should be explicit and not default to some directory.

Avoid this default directory gets rid of this startup message for
everyone not using callbacks:
DEBUG - Callback configuration path doesn't exist ./callbacks

Instead we can now exit with an error if the specified callbacks
directory does not exist, instead of just noting it at DEBUG level.

Also rename --callback-path to --callback-dir to harmonize with
--specs-dir and --log-dir naming.