Adds scope for server
Scope is a way to signal the current scope of the test.
For example, if we are testing A and then B, we set
the scope to A and then B, and by querying the scope, we can
assign API requests and responses to their logical "place"
in things like reports.