Serves back correct status code. (#152)
* Serves back correct status code.

This required deleting one obsolete test that only passed because
of the bug but then started to fail. As the test itself was a bit
ambiguous (not clear what it was supposed to accomplish) and
as a similar test existed in another file, I deleted it. The Stripe
test also failed, revealing two interesting bugs in the faker that
got fixed as well.